+++ packages/components/src/components/index.ts
@@ -1,2 +1,2 @@
-export * from './layout';
+export * from './layouts';
 export * from './primitives';
