+++ packages/components/src/components/layouts/Section/Section.tsx
@@ -1,6 +1,6 @@
 import { Box } from '@chakra-ui/layout';
 import React, { FC } from 'react';
-import { SectionProps } from './';
+import { SectionProps } from '.';
 
 export const Section: FC<SectionProps> = (props) => (
   <Box {...props} as="section" />
